Add bitmapReadyFlow to signal bitmap is fetched and break onPageUpdate with granular events
This CL introduces a new SharedFlow called bitmapReadyFlow in PageManager. This flow emits the page number whenever new bitmaps for that page are ready. It also refactors the page update mechanism by replacing the onPageUpdate callback with more specific onBitmapReady and onFormWidgetReady callbacks.

- `Page` and `BitmapFetcher` now use `onBitmapReady(Int)` when a bitmap is rendered and `onFormWidgetReady(Int)` when form widgets are loaded.
- PageManager introduces a new bitmapReadyFlow to signal PdfView when page content is ready.

Test: Build and run on local device
Bug: 461568127

+++ pdf/pdf-viewer/src/test/kotlin/androidx/pdf/view/BitmapFetcherTest.kt
@@ -48,7 +48,7 @@ class BitmapFetcherTest {
         }
 
     private var invalidationCounter = 0
-    private val invalidationTracker: () -> Unit = { invalidationCounter++ }
+    private val invalidationTracker: (Int) -> Unit = { invalidationCounter++ }
 
     private val maxBitmapSizePx = Point(2048, 2048)
     private val pageSize = Point(512, 512)
